requirements.txt
streamlit==1.44.1
python-dotenv==1.0.1
langchain-core==0.3.56
langchain-community==0.3.23
langchain-google-genai==2.1.3
langchain-text-splitters==0.3.8
faiss-cpu==1.8.0.post1
unstructured==0.16.23
pypdf==5.4.0
